pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mattermost-tui"
version = "0.1.0"
description = "Terminal UI for Mattermost (REST login and chat)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "httpx[socks]>=0.27.0",
    "python-dotenv>=1.0.0",
    "textual>=0.86.0",
]

[project.urls]
Repository = "https://github.com/aaronsdevera/mattermost-tui"

[project.scripts]
mattermost-tui = "mattermost_tui.cli:main"
mmt = "mattermost_tui.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["mattermost_tui"]

[dependency-groups]
build-binary = ["pyinstaller>=6.11"]
dev = ["pytest>=8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = ["-q"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_date_format = "%H:%M:%S"
log_level = "DEBUG"
log_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"

[[tool.uv.index]]
url = "https://pypi.org/simple/"
